+++ src/mcp_server/services/datainclusion/service.py
@@ -8,14 +8,13 @@ demandées pour rendre certains paramètres de recherche obligatoires.
 import os
 import logging
 import httpx
-import functools
 from typing import List, Optional, Union, Literal
 
 from fastmcp import FastMCP
 from fastmcp.tools import Tool
-from pydantic_ai import ModelRetry
 from starlette.requests import Request
 from starlette.responses import PlainTextResponse
+from ..utils import api_call_handler
 from .schemas import (
     ReferenceItem,
     StructureSummary,
@@ -81,33 +80,6 @@ _REFERENCE_ENDPOINTS = {
     "networks": "/api/v1/doc/reseaux-porteurs",
     "service_types": "/api/v1/doc/types-services",
 }
-
-
-def api_call_handler(func):
-    """Décorateur pour la gestion centralisée des appels API, du logging et des erreurs."""
-
-    @functools.wraps(func)
-    async def wrapper(*args, **kwargs):
-        tool_name = func.__name__
-        try:
-            return await func(*args, **kwargs)
-        except httpx.HTTPStatusError as e:
-            logger.error(
-                f"Erreur HTTP dans '{tool_name}': {e.response.status_code} - {e.response.text}"
-            )
-            raise ModelRetry(
-                f"Erreur de communication avec l'API Data Inclusion ({e.response.status_code}): {e.response.text}"
-            ) from e
-        except httpx.HTTPError as e:
-            logger.error(f"Erreur de communication dans '{tool_name}': {e}")
-            raise ModelRetry(
-                f"Erreur de communication avec l'API Data Inclusion: {e}"
-            ) from e
-        except Exception as e:
-            logger.error(f"Erreur inattendue dans '{tool_name}': {e}", exc_info=True)
-            raise ModelRetry(f"Une erreur inattendue s'est produite: {e}") from e
-
-    return wrapper
 
 
 # --- Définition des Outils ---
